backend/requirements.txt
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dlib==20.0.0
drf-spectacular==0.28.0
face-recognition==1.3.0
face_recognition_models==0.3.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
numpy==2.3.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.8.0
pyOpenSSL==25.3.0
pytz==2024.1
PyYAML==6.0.3
redis==7.0.1
referencing==0.37.0
rpds-py==0.28.0
service-identity==24.2.0
sqlparse==0.5.0
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
zope.interface==8.0.1
